improvements: added global handlers' for api errors
