+++ src/components/group/EditGroupNameModal.tsx
@@ -17,7 +17,7 @@ import { useRouter } from 'next/router'
 import React, { useEffect, useState } from 'react'
 
 import { editGroup } from '@/utils/api_helper'
-import { isJson } from '@/utils/generic_utils'
+import { handleApiError } from '@/utils/error_helper'
 
 export interface IEditGroupNameModal {
   groupID: string | null
@@ -32,8 +32,6 @@ const EditGroupNameModal: React.FC<IEditGroupNameModal> = props => {
   const [input, setInput] = useState('')
   const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) =>
     setInput(e.target.value)
-  const toast = useToast()
-  const { push } = useRouter()
   useEffect(() => {
     if (props.groupName) {
       setInput(props.groupName)
@@ -49,18 +47,7 @@ const EditGroupNameModal: React.FC<IEditGroupNameModal> = props => {
       props.resetState()
       props.onClose()
     } catch (error: any) {
-      const isJsonErr = isJson(error.message)
-      const errorMessage = isJsonErr
-        ? JSON.parse(error.message)?.error
-        : error.message
-      toast({
-        title: 'Error changing name',
-        description: errorMessage,
-        status: 'error',
-        duration: 5000,
-        isClosable: true,
-        position: 'top',
-      })
+      handleApiError('Error changing name', error)
     }
     setIsUpdating(false)
   }
