+++ src/components/group/GroupInviteForm.tsx
@@ -17,7 +17,7 @@ import InfoTooltip from '@/components/profile/components/Tooltip'
 import { GroupInvitePayload, MemberType } from '@/types/Group'
 import { InvitedUser } from '@/types/ParticipantInfo'
 import { getExistingAccounts, inviteUsers } from '@/utils/api_helper'
-import { isJson } from '@/utils/generic_utils'
+import { handleApiError } from '@/utils/error_helper'
 import {
   isEthereumAddressOrDomain,
   isValidEmail,
@@ -66,7 +66,7 @@ const GroupInviteForm: FC<InviteModalProps> = ({
 
     const payload: GroupInvitePayload = { invitees, message }
     try {
-      const resp = await inviteUsers(groupId, payload)
+      await inviteUsers(groupId, payload)
 
       toast({
         title: 'Invitation sent successfully',
@@ -89,18 +89,7 @@ const GroupInviteForm: FC<InviteModalProps> = ({
       setInvitedUsers([])
       onInviteSuccess?.()
     } catch (error: any) {
-      const isJsonErr = isJson(error.message)
-      const errorMessage = isJsonErr
-        ? JSON.parse(error.message)?.error
-        : error.message
-      toast({
-        title: 'Error inviting member',
-        description: errorMessage,
-        status: 'error',
-        duration: 5000,
-        isClosable: true,
-        position: 'top',
-      })
+      handleApiError('Error inviting member', error)
     }
     setIsSaving(false)
   }
